(PC-17814) identityCheck: move errors.ts file to /profile folder
